requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
cloudinary==1.26.0
coreapi==2.3.3
coreschema==0.0.4
dateparser==1.0.0
discord.py==1.7.3
distlib==0.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.6
django-cloudinary-storage==0.3.0
django-cors-headers==3.8.0
django-dotenv==1.4.2
django-filter==21.1
djangorestframework==3.12.4
drf-yasg==1.20.0
filelock==3.0.12
gunicorn==20.1.0
html5lib==1.1
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
jsonschema==4.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
multidict==5.1.0
packaging==21.0
Pillow==8.3.1
psycopg==3.0b1
psycopg2==2.9.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
regex==2021.8.28
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
static3==0.7.0
swagger-spec-validator==2.7.3
typing-extensions==3.10.0.0
tzdata==2021.1
tzlocal==3.0
uritemplate==4.1.1
urllib3==1.26.6
virtualenv==20.4.6
webencodings==0.5.1
whitenoise==5.3.0
xmltodict==0.12.0
yarl==1.6.3
